separate cache evaluation and 2dto transformation
Introducing a Cache with lazy creation support allows to move
the cache evaluation out of the scope of the transformation and into the
the scope controllers.

Doing so pointed out, that for single dto responses an empty cache was
used, i.e. the code would look like

    new Cache.get(key).orElseCreate(() -> transform(entity))

wich will always use the supplier `() -> transform(entity)`.

Fro list dto responses the cache would have an effect iff the list
contains the same element multiple twice, which would actually be a bug.

As a conclusion the cache will not be used anymore.

VEO-342