Disable autocommit check
By default Hibernate checks and turns off autocommit for every connection. Let Hikari do it instead and tell Hibernate that it doesn't need to check every connection. This allows connections to be acquired when they're actually needed to execute a statement instead of at the beginnign of a transaction to check its autocommit status.