tag default docker image with git hash
The Jenkins docker plugin uses the docker image name for consequent
steps, which causes race conditions when other branches are build in
parallel. E.g. consider the master and develop branch beeing build in
parallel

    docker build -t eu.gcr.io/veo-projekt/veo . # building master
    docker build -t eu.gcr.io/veo-projekt/veo . # building develop

Since so tag is specified docker assumes both are tagged as *latest*.
This means in this first master is tagged as latest but then develop is
tagged as latest in the local docker storage, too. Which means the job
building the master branch will now push the development build as latest
when running

    docker tag eu.gcr.io/veo-projekt/veo eu.gcr.io/veo-projekt/veo:latest
    docker push eu.gcr.io/veo-projekt/veo:latest