Fix domain ownership
This commit fixes "CreateUnitUseCaseSpec.groovy" and the corresponding
use case: a domain has to be added to the client when a new unit is
created. Otherwise, no domain is available when a unit or any entity is
created. Domains are owned by the client and only referenced from other
entities. Existing code did not honor this rule. Other existing tests
did not care about no domain being present.

It is currently impossible to create a domain (required when creating a
new client) using the existing REST endpoints. The existing tests only
worked because they create a domain object using the domain repository
directly. It is also impossible to add or remove domains using the REST
endpoints. This commit makes the problem visible, fixes tests and
contains a woraround for VEO-227.  VEO-227